OCR-tests/requirements.txt
pytesseract
paddlepaddle
paddleocr
jiwer
opencv-python
pandas
transformers
torch
Pillow